Rename util.js into generator.js
